ENH: Generate the sample list using dynamic threshold
A new feature allows generating the sample list using a dynamic
threshold. A threshold is calculated around the pixel using the
IntensityBinSize and min and max values of threshold. A connected
threshold is computed and the size of the patch is calculated.
A listSample is generated and passed to the histogram generator and
then to the run length features calculation.